requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-docx==1.1.0
httpx==0.26.0
anthropic==0.18.1
python-dotenv==1.0.0
itsdangerous==2.1.2
